Move `RuntimeParameter` to `step/typing.py`
