+++ tests/integration/test_pipe_llms.py
@@ -18,10 +18,10 @@ from typing import Any, Dict, Generator, List
 from distilabel.llm.huggingface.transformers import TransformersLLM
 from distilabel.llm.openai import OpenAILLM
 from distilabel.pipeline.local import Pipeline
-from distilabel.pipeline.step.base import RuntimeParameter, Step
+from distilabel.pipeline.step.base import Step
 from distilabel.pipeline.step.generators.huggingface import LoadHubDataset
 from distilabel.pipeline.step.task.generation import TextGeneration
-from distilabel.pipeline.step.typing import StepInput
+from distilabel.pipeline.step.typing import RuntimeParameter, StepInput
 
 
 class RenameColumns(Step):
@@ -33,13 +33,13 @@ class RenameColumns(Step):
 
     @property
     def outputs(self) -> List[str]:
-        return list(self.rename_mappings.values())
+        return list(self.rename_mappings.values())  # type: ignore
 
     def process(self, inputs: StepInput) -> Generator[List[Dict[str, Any]], None, None]:
         outputs = []
         for input in inputs:
             outputs.append(
-                {self.rename_mappings.get(k, k): v for k, v in input.items()}
+                {self.rename_mappings.get(k, k): v for k, v in input.items()}  # type: ignore
             )
         yield outputs
