+++ tests/unit/pipeline/test_dag.py
@@ -19,8 +19,8 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, List
 import pytest
 from distilabel.pipeline._dag import DAG
 from distilabel.pipeline.local import Pipeline
-from distilabel.pipeline.step.base import GeneratorStep, RuntimeParameter, Step
-from distilabel.pipeline.step.typing import StepInput
+from distilabel.pipeline.step.base import GeneratorStep, Step
+from distilabel.pipeline.step.typing import RuntimeParameter, StepInput
 
 from .utils import DummyGeneratorStep
